requirements.txt
streamlit
pickle
numpy
pillow
dotenv
huggingface_hub